console: Refactor gateway title section
